move Widget and IsWidget wrapper to its own module
