+++ core/src/main/java/org/jboss/gwt/elemento/core/Elements.java
@@ -24,8 +24,6 @@ import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
 import com.google.gwt.safehtml.shared.SafeHtml;
-import com.google.gwt.user.client.ui.IsWidget;
-import com.google.gwt.user.client.ui.Widget;
 import elemental2.dom.Element;
 import elemental2.dom.HTMLAnchorElement;
 import elemental2.dom.HTMLAreaElement;
@@ -731,36 +729,6 @@ public final class Elements {
         return "elemento-uid-" + createDocumentUniqueId.getAsInt();
     }
 
-
-    // ------------------------------------------------------ conversions
-
-
-    /** Converts from {@link IsElement} &rarr; {@link Widget}. */
-    public static Widget asWidget(IsElement element) {
-        return asWidget(element.asElement());
-    }
-
-    /** Converts from {@link HTMLElement} &rarr; {@link Widget}. */
-    public static Widget asWidget(HTMLElement element) {
-        return new ElementWidget(element);
-    }
-
-    /** Converts from {@link IsWidget} &rarr; {@link HTMLElement}. */
-    public static HTMLElement asElement(IsWidget widget) {
-        return asElement(widget.asWidget());
-    }
-
-    /** Converts from {@link Widget} &rarr; {@link HTMLElement}. */
-    public static HTMLElement asElement(Widget widget) {
-        return asElement(widget.getElement());
-    }
-
-    /** Converts from {@link com.google.gwt.dom.client.Element} &rarr; {@link HTMLElement}. */
-    public static HTMLElement asElement(com.google.gwt.dom.client.Element element) {
-        return Js.cast(element);
-    }
-
-
     // this is a static helper class which must never be instantiated!
     private Elements() {
 
@@ -777,15 +745,6 @@ public final class Elements {
             return ((HTMLElement) t);
         }
     }
-
-
-    private static class ElementWidget extends Widget {
-
-        ElementWidget(HTMLElement element) {
-            setElement(com.google.gwt.dom.client.Element.as(Js.cast(element)));
-        }
-    }
-
 
     private static class FilterHTMLElements<T extends Node> implements Predicate<T> {
